changed return type

+++ src/execution/riscv32/simulator-riscv32.h
@@ -108,14 +108,14 @@ static inline bool isSnan(float fp) { return !QUIET_BIT_S(fp); }
 static inline bool isSnan(double fp) { return !QUIET_BIT_D(fp); }
 #undef QUIET_BIT_S
 #undef QUIET_BIT_D
-// RV32Gtodo: in RV32 mul's operands are 32bit
-inline uint64_t mulhu(uint32_t a, uint32_t b) {
+
+inline uint32_t mulhu(uint32_t a, uint32_t b) {
   uint64_t full_result = ((uint64_t)a) * ((uint64_t)b);
   uint32_t upper_part = full_result >> 32;
   return upper_part;
 }
 
-inline int64_t mulh(int32_t a, int32_t b) {
+inline int32_t mulh(int32_t a, int32_t b) {
   int64_t full_result = ((int64_t)a) * ((int64_t)b);
   int32_t upper_part = full_result >> 32;
   return upper_part;
